+++ stac/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - `Links::remove_structural_links`, and more rel types to `Link::is_structural` ([#170](https://github.com/gadomski/stac-rs/pull/170))
 
+### Removed
+
+- `Link::set_query` ([#171](https://github.com/gadomski/stac-rs/pull/171))
+
 ## [0.4.0] - 2023-04-01
 
 ### Added